refactor(deck): 全カードの生成はdeckのコンストラクタ内で実装する
