+++ src/main.rs
@@ -4,6 +4,7 @@ use blackjack::card_group::Deck;
 
 fn main() {
     let mut deck = Deck::new();
+    deck.shuffle();
     println!("{}", deck);
     let top = deck.draw();
     println!("{}", top.unwrap());
